+++ bika/lims/browser/analysisrequest/add2.py
@@ -1013,16 +1013,24 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
         })
 
         # catalog queries for UI field filtering
+        sample_type_uid = api.get_uid(obj)
         filter_queries = {
+            # Display Sample Points that have this sample type assigned plus
+            # those that do not have a sample type assigned
             "SamplePoint": {
-                "getSampleTypeTitles": [obj.Title(), ''],
+                "sampletype_uids": [sample_type_uid, None],
                 "getClientUID": [client_uid, ""],
-                "sort_order": "descending",
             },
+            # Display Specifications that have this sample type assigned only
             "Specification": {
-                "getSampleTypeTitle": obj.Title(),
+                "sampletype_uids": sample_type_uid,
                 "getClientUID": [client_uid, ""],
-                "sort_order": "descending",
+            },
+            # Display AR Templates that have this sample type assigned plus
+            # those that do not have a sample type assigned
+            "Template": {
+                "sampletype_uids": [sample_type_uid, None],
+                "getClientUID": [client_uid, ""],
             }
         }
         info["filter_queries"] = filter_queries
@@ -1152,7 +1160,8 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             ],
             "SampleType": [
                 "SamplePoint",
-                "Specification"
+                "Specification",
+                "Template",
             ],
             "PrimarySample": [
                 "Batch"