+++ bika/lims/content/samplepoint.py
@@ -25,7 +25,6 @@ from Products.ATContentTypes.lib.historyaware import HistoryAwareMixin
 from Products.Archetypes.public import BaseContent
 from Products.Archetypes.public import BooleanField
 from Products.Archetypes.public import BooleanWidget
-from Products.Archetypes.public import DisplayList
 from Products.Archetypes.public import FileWidget
 from Products.Archetypes.public import ReferenceField
 from Products.Archetypes.public import Schema
@@ -34,8 +33,10 @@ from Products.Archetypes.public import StringWidget
 from Products.Archetypes.public import registerType
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_unicode
+from plone.app.blob.field import FileField as BlobFileField
+from zope.interface import implements
+
 from bika.lims import bikaMessageFactory as _
-from bika.lims import deprecated
 from bika.lims.browser.fields import CoordinateField
 from bika.lims.browser.fields import DurationField
 from bika.lims.browser.widgets import CoordinateWidget
@@ -45,9 +46,8 @@ from bika.lims.browser.widgets.referencewidget import \
 from bika.lims.config import PROJECTNAME
 from bika.lims.content.bikaschema import BikaSchema
 from bika.lims.content.clientawaremixin import ClientAwareMixin
+from bika.lims.content.sampletype import SampleTypeAwareMixin
 from bika.lims.interfaces import IDeactivable
-from plone.app.blob.field import FileField as BlobFileField
-from zope.interface import implements
 
 schema = BikaSchema.copy() + Schema((
     CoordinateField(
@@ -91,13 +91,17 @@ schema = BikaSchema.copy() + Schema((
         required=0,
         multiValued=1,
         allowed_types=('SampleType',),
-        vocabulary='SampleTypesVocabulary',
         relationship='SamplePointSampleType',
         widget=BikaReferenceWidget(
             label=_("Sample Types"),
             description=_("The list of sample types that can be collected "
                           "at this sample point.  If no sample types are "
                           "selected, then all sample types are available."),
+            catalog_name='bika_setup_catalog',
+            base_query={"is_active": True,
+                        "sort_on": "sortable_title",
+                        "sort_order": "ascending"},
+            showOn=True,
         ),
     ),
 
@@ -126,7 +130,8 @@ schema['description'].widget.visible = True
 schema['description'].schemata = 'default'
 
 
-class SamplePoint(BaseContent, HistoryAwareMixin, ClientAwareMixin):
+class SamplePoint(BaseContent, HistoryAwareMixin, ClientAwareMixin,
+                  SampleTypeAwareMixin):
     implements(IDeactivable)
     security = ClassSecurityInfo()
     displayContentsTab = False
@@ -140,31 +145,6 @@ class SamplePoint(BaseContent, HistoryAwareMixin, ClientAwareMixin):
 
     def Title(self):
         return safe_unicode(self.getField('title').get(self)).encode('utf-8')
-
-    def getSampleTypeTitles(self):
-        """Returns a list of sample type titles
-        """
-        sample_types = self.getSampleTypes()
-        sample_type_titles = map(lambda obj: obj.Title(), sample_types)
-
-        # N.B. This is used only for search purpose, because the catalog does
-        #      not add an entry to the Keywordindex for an empty list.
-        #
-        # => This "empty" category allows to search for values with a certain
-        #    sample type set OR with no sample type set.
-        #    (see bika.lims.browser.analysisrequest.add2.get_sampletype_info)
-        if not sample_type_titles:
-            return [""]
-        return sample_type_titles
-
-    def getSampleTypeTitle(self):
-        """Returns a comma separated list of sample type titles
-        """
-        return ",".join(self.getSampleTypeTitles())
-
-    def SampleTypesVocabulary(self):
-        from bika.lims.content.sampletype import SampleTypes
-        return SampleTypes(self, allow_blank=False)
 
     def setSampleTypes(self, value, **kw):
         """ For the moment, we're manually trimming the sampletype<>samplepoint
@@ -200,31 +180,5 @@ class SamplePoint(BaseContent, HistoryAwareMixin, ClientAwareMixin):
 
         return ret
 
-    def getSampleTypes(self, **kw):
-        return self.Schema()['SampleTypes'].get(self)
-
 
 registerType(SamplePoint, PROJECTNAME)
-
-
-@deprecated("bika.lims.content.samplepoint.SamplePoints function will be removed in senaite.core 1.2.0")
-def SamplePoints(self, instance=None, allow_blank=True, lab_only=True):
-    instance = instance or self
-    bsc = getToolByName(instance, 'bika_setup_catalog')
-    items = []
-    contentFilter = {
-        'portal_type': 'SamplePoint',
-        'is_active': True,
-        'sort_on': 'sortable_title'}
-    if lab_only:
-        lab_path = instance.bika_setup.bika_samplepoints.getPhysicalPath()
-        contentFilter['path'] = {"query": "/".join(lab_path), "level": 0}
-    for sp in bsc(contentFilter):
-        sp = sp.getObject()
-        if sp.aq_parent.portal_type == 'Client':
-            sp_title = "{}: {}".format(sp.aq_parent.Title(), sp.Title())
-        else:
-            sp_title = sp.Title()
-        items.append((sp.UID(), sp_title))
-    items = allow_blank and [['', '']] + list(items) or list(items)
-    return DisplayList(items)